Move fork router functions to a standalone file
